+++ project/OsEngine/Market/Servers/Bitfinex/BitfinexSpot/BitfinexSpotServer.cs
@@ -2029,16 +2029,16 @@ namespace OsEngine.Market.Servers.Bitfinex
                     {
                         bids.Add(new MarketDepthLevel()
                         {
-                            Bid = (value[2]).ToString().ToDecimal(),
-                            Price = (value[0]).ToString().ToDecimal(),
+                            Bid = (value[2]).ToString().ToDouble(),
+                            Price = (value[0]).ToString().ToDouble(),
                         });
                     }
                     else
                     {
                         asks.Add(new MarketDepthLevel()
                         {
-                            Ask = Math.Abs((value[2]).ToString().ToDecimal()),
-                            Price = (value[0]).ToString().ToDecimal(),
+                            Ask = Math.Abs((value[2]).ToString().ToDouble()),
+                            Price = (value[0]).ToString().ToDouble(),
                         });
                     }
                 }
@@ -2112,9 +2112,9 @@ namespace OsEngine.Market.Servers.Bitfinex
                     return;
                 }
 
-                decimal price = (update[0]).ToString().ToDecimal();
-                decimal count = (update[1]).ToString().ToDecimal();
-                decimal amount = (update[2]).ToString().ToDecimal();
+                double price = (update[0]).ToString().ToDouble();
+                double count = (update[1]).ToString().ToDouble();
+                double amount = (update[2]).ToString().ToDouble();
 
                 needDepth.Time = ServerTime;
